42_tran/be/requirements.txt
Django>=3.0,<4.0
psycopg2-binary
hvac
djangorestframework>=3.12.0
django-allauth>=0.44.0
dj-rest-auth>=2.1.0
channels>=3.0.0
